feat: Add classification and classification_reason
AWS has added seeing the classification (and reason) on an HTTP request, to see
if it's trying to bypass HTTP Desync mitigations AWS has put on. This allows us
to classify these fields.

We specifically put PLAIN_DICTIONARY on both Classification and
ClassificationReason because of the limited number of variants that can go in
this field.