test: Mirror new parsing test cases
With the addition of the new fields, we have two changes here:

1. Shifting the OtherFields to the very last case yet again, to accommodate the
four extra fields newly added. If it's kept on the old test case, it would be
erroneously classified as not OtherFields.
2. Adding the new entries added in data.log to the parsing test cases. This also
includes an entry that has four extra fields from the previous last test case,
and moves it to the end.